+++ Project1/Line2D.h
@@ -5,7 +5,7 @@ class Line2D
 {
 public:
 	Line2D(const Point2D& p1 = Point2D(), const Point2D& p2 = Point2D());
-	Line2D(const Point2D& p, float direction, float width); // direction in degree
+	Line2D(const Point2D& p, double direction, double width); // direction in degree
 
 	Point2D getPoint1() const;
 	Point2D getPoint2() const;
@@ -13,13 +13,13 @@ public:
 	void setPoint1(const Point2D& p);
 	void setPoint2(const Point2D& p);
 
-	float getLength() const; // in meters.
-	float distanceToLine(const Point2D& p) const; // in meters.
-	float crossValue(const Point2D& p) const;
+	double getLength() const; // in meters.
+	double distanceToLine(const Point2D& p) const; // in meters.
+	double crossValue(const Point2D& p) const;
 	bool isPointInInterval(const Point2D& p) const;
 
 private:
 	Point2D point1;
 	Point2D point2;
-	float length;
+	double length;
 };
